Before authentication in ideas-api project

+++ apps/ideas-api/src/app/idea/idea.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common';
+import { Injectable, NotFoundException } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
 import { IdeaDTO } from './idea.dto';
@@ -8,30 +8,48 @@ import { IdeaEntity } from './idea.entity';
 export class IdeaService {
   constructor(
     @InjectRepository(IdeaEntity)
-    private ideaRepository: Repository<IdeaEntity>
-    ) {}
+    private ideaRepository: Repository<IdeaEntity>,
+  ) {}
 
-    async showAll() {
-      return await this.ideaRepository.find();
+  async showAll() {
+    return await this.ideaRepository.find();
+  }
+
+  async create(data: IdeaDTO) {
+    const idea = this.ideaRepository.create(data);
+    await this.ideaRepository.save(idea);
+    return idea;
+  }
+
+  async read(id: string) {
+    const idea = await this.ideaRepository.findOne({ where: { id } });
+    if (!idea) {
+      throw new NotFoundException();
+    }
+    console.log(idea);
+
+    return idea;
+  }
+
+  async update(id: string, data: Partial<IdeaDTO>) {
+    const idea = await this.ideaRepository.findOne({where: {id}});
+
+    if(!idea) {
+      throw new NotFoundException();
+    }
+    await this.ideaRepository.update({ id }, data);
+    return idea;
+  }
+
+  async destroy(id: string) {
+
+    const idea = await this.ideaRepository.findOne({where: {id}});
+
+    if(!idea) {
+      throw new NotFoundException();
     }
 
-    async create(data: IdeaDTO) {
-      const idea = this.ideaRepository.create(data);
-      await this.ideaRepository.save(idea);
-      return idea;
-    }
-
-    async read(id: string) {
-      return this.ideaRepository.findOne({where: {id}})
-    }
-
-    async update(id: string, data: Partial<IdeaDTO>) {
-      await this.ideaRepository.update({id}, data);
-      return await this.ideaRepository.findOne({id});
-    }
-
-    async destroy(id: string) {
-      await this.ideaRepository.delete({id});
-      return {deleted: true};
-    }
+    await this.ideaRepository.delete({ id });
+    return { deleted: true };
+  }
 }
